requirements-drive-sync.txt
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.122.0
